Fixed an issue with group_paginate function removing values when used with a bool dimension.
